#4238 Provide more scanning context for batch aggregation

+++ processing/src/main/java/io/druid/segment/serde/DoubleGenericColumnPartSerde.java
@@ -126,7 +126,7 @@ public class DoubleGenericColumnPartSerde implements ColumnPartSerde
                      Integer.BYTES +  // elements num
                      Integer.BYTES +  // sizePer
                      1 +              // compression id
-                     Double.BYTES * numRows;
+                     (long) Double.BYTES * numRows;
             }
 
             @Override
